requirements.txt
mysql-connector-python>=8.0.30
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
openpyxl
requests
python-dateutil
fastapi
uvicorn